Refactor postgres migration generation
