+++ lib/ecto/postgres/pg_rational_migrations.ex
@@ -5,17 +5,15 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Migrations for adding rational types, casts, functions and constraints to a
   Postgres database.
   """
+  use Vtc.Ecto.Postgres.PgTypeMigration
+
   alias Ecto.Migration
   alias Ecto.Migration.Constraint
-  alias Vtc.Ecto.Postgres
   alias Vtc.Ecto.Postgres.Fragments
+  alias Vtc.Ecto.Postgres.PgRational
+  alias Vtc.Ecto.Postgres.PgTypeMigration
 
   require Ecto.Migration
-
-  @typedoc """
-  Indicates returned string is an SQL command.
-  """
-  @type raw_sql() :: String.t()
 
   @doc section: :migrations_full
   @doc """
@@ -86,29 +84,6 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   - `functions_prefix`: A prefix to add before all functions. Defaults to "rational"
     for any function created in the `:public` schema, and "" otherwise.
 
-  ## Private Functions
-
-  Some custom function names are prefaced with `__private__`. These functions should
-  not be called by end-users, as they are not subject to *any* API stability guarantees.
-
-  ## Functions Created
-
-  See [PgFunctions](Vtc.Ecto.Postgres.PgRational.Migrations.html#pgfunctions)
-  section of these docs for details on native database functions
-  created.
-
-  ## Operators Created
-
-  See [PgOperators](Vtc.Ecto.Postgres.PgRational.Migrations.html#pgoperators)
-  section of these docs for details on native database operators
-  created.
-
-  ## Casts Created
-
-  See [PgCasts](Vtc.Ecto.Postgres.PgRational.Migrations.html#pgcasts)
-  section of these docs for details on native database operators
-  created.
-
   ## Examples
 
   ```elixir
@@ -119,14 +94,22 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
     require PgRational.Migrations
 
     def change do
-      PgRational.Migrations.create_all()
+      PgRational.Migrations.run()
     end
   end
   ```
   """
-  @spec create_all(include: Keyword.t(), exclude: Keyword.t()) :: :ok
-  def create_all(opts \\ []) do
-    migrations = [
+  @spec run(include: Keyword.t(atom()), exclude: Keyword.t(atom())) :: :ok
+  def run(opts \\ []), do: PgTypeMigration.run_for(__MODULE__, opts)
+
+  @doc false
+  @impl PgTypeMigration
+  def ecto_type, do: PgRational
+
+  @doc false
+  @impl PgTypeMigration
+  def migrations_list do
+    [
       &create_type/0,
       &create_function_schemas/0,
       &create_func_simplify/0,
@@ -167,8 +150,6 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
       &create_cast_double_precision/0,
       &create_cast_bigint_to_rational/0
     ]
-
-    Postgres.Utils.run_migrations(migrations, opts)
   end
 
   @doc section: :migrations_types
@@ -179,9 +160,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   - `rationals` schema
   - `rationals_helpers` schema
   """
-  @spec create_type() :: {raw_sql(), raw_sql()}
+  @spec create_type() :: migration_info()
   def create_type do
-    Postgres.Utils.create_type(:rational,
+    PgTypeMigration.create_type(:rational,
       numerator: :bigint,
       denominator: :bigint
     )
@@ -193,17 +174,17 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   [Configuring Database Objects](Vtc.Ecto.Postgres.PgRational.Migrations.html#create_all/0-configuring-database-objects)
   section above.
   """
-  @spec create_function_schemas() :: {raw_sql(), raw_sql()}
-  def create_function_schemas, do: Postgres.Utils.create_type_schema(:rational)
+  @spec create_function_schemas() :: migration_info()
+  def create_function_schemas, do: PgTypeMigration.create_type_schema(:rational)
 
   @doc section: :migrations_private_functions
   @doc """
   Creates `rational.__private__simplify(rat)` function that simplifies a rational. Used at
   the end of every rational operation to avoid overflows.
   """
-  @spec create_func_simplify() :: {raw_sql(), raw_sql()}
+  @spec create_func_simplify() :: migration_info()
   def create_func_simplify do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:simplify, Migration.repo()),
       args: [input: :rational],
       returns: :rational,
@@ -224,9 +205,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `rational.__private__minus(rat)` function that flips the sign of the input
   value -- makes a positive value negative and a negative value positive.
   """
-  @spec create_func_minus() :: {raw_sql(), raw_sql()}
+  @spec create_func_minus() :: migration_info()
   def create_func_minus do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:minus, Migration.repo()),
       args: [input: :rational],
       returns: :rational,
@@ -241,9 +222,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `ABS(rational)` function that returns the absolute value of the rational
   value.
   """
-  @spec create_func_abs() :: {raw_sql(), raw_sql()}
+  @spec create_func_abs() :: migration_info()
   def create_func_abs do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       "ABS",
       args: [input: :rational],
       returns: :rational,
@@ -258,9 +239,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `ABS(rational)` function that returns the absolute value of the rational
   value.
   """
-  @spec create_func_sign() :: {raw_sql(), raw_sql()}
+  @spec create_func_sign() :: migration_info()
   def create_func_sign do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       "SIGN",
       args: [input: :rational],
       returns: :integer,
@@ -275,9 +256,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `ROUND(rational)` function that returns the rational input, rounded to the
   nearest :bigint.
   """
-  @spec create_func_round() :: {raw_sql(), raw_sql()}
+  @spec create_func_round() :: migration_info()
   def create_func_round do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       "ROUND",
       args: [input: :rational],
       returns: :bigint,
@@ -300,9 +281,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `FLOOR(rational)` function that returns the rational input as a `bigint`,
   rounded towards zero, to match Postgres `FLOOR(real)` behavior.
   """
-  @spec create_func_floor() :: {raw_sql(), raw_sql()}
+  @spec create_func_floor() :: migration_info()
   def create_func_floor do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       "FLOOR",
       args: [input: :rational],
       returns: :bigint,
@@ -316,9 +297,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a native CAST from `rational` to `double precision`.
   """
-  @spec create_func_cast_to_double_precision() :: {raw_sql(), raw_sql()}
+  @spec create_func_cast_to_double_precision() :: migration_info()
   def create_func_cast_to_double_precision do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:cast_to_double, Migration.repo()),
       args: [value: :rational],
       returns: :"double precision",
@@ -332,9 +313,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a native CAST from `bigint` to `rational`.
   """
-  @spec create_func_cast_bigint_to_rational() :: {raw_sql(), raw_sql()}
+  @spec create_func_cast_bigint_to_rational() :: migration_info()
   def create_func_cast_bigint_to_rational do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:cast_bigint_to_rational, Migration.repo()),
       args: [value: :bigint],
       returns: :rational,
@@ -351,9 +332,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `rational.__private__add(a, b)` backing function for the `+` operator
   between two rationals.
   """
-  @spec create_func_add() :: {raw_sql(), raw_sql()}
+  @spec create_func_add() :: migration_info()
   def create_func_add do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:add, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: [
@@ -375,9 +356,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `rational.__private__sub(a, b)` backing function for the `-` operator
   between two rationals.
   """
-  @spec create_func_sub() :: {raw_sql(), raw_sql()}
+  @spec create_func_sub() :: migration_info()
   def create_func_sub do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:sub, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: [
@@ -399,9 +380,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `rational.__private__mult(a, b)` backing function for the `*` operator
   between two rationals.
   """
-  @spec create_func_mult() :: {raw_sql(), raw_sql()}
+  @spec create_func_mult() :: migration_info()
   def create_func_mult do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:mult, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: [
@@ -423,9 +404,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `rational.__private__div(a, b)` backing function for the `/` operator
   between two rationals.
   """
-  @spec create_func_div() :: {raw_sql(), raw_sql()}
+  @spec create_func_div() :: migration_info()
   def create_func_div do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:div, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: [
@@ -449,9 +430,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
 
   Just like `DIV(real, real)`, `DIV(rational, rational)` floors towards zero.
   """
-  @spec create_func_floor_div() :: {raw_sql(), raw_sql()}
+  @spec create_func_floor_div() :: migration_info()
   def create_func_floor_div do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       "DIV",
       args: [a: :rational, b: :rational],
       declares: [
@@ -470,9 +451,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   Creates `rational.__private__modulo(a, b)` backing function for the `%` operator
   between two rationals.
   """
-  @spec create_func_modulo() :: {raw_sql(), raw_sql()}
+  @spec create_func_modulo() :: migration_info()
   def create_func_modulo do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:modulo, Migration.repo()),
       args: [dividend: :rational, divisor: :rational],
       declares: [
@@ -506,9 +487,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
 
   Used to back equality operators.
   """
-  @spec create_func_cmp() :: {raw_sql(), raw_sql()}
+  @spec create_func_cmp() :: migration_info()
   def create_func_cmp do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:cmp, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: [
@@ -526,9 +507,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates `rational.__private__eq(a, b)` that backs the `=` operator.
   """
-  @spec create_func_eq() :: {raw_sql(), raw_sql()}
+  @spec create_func_eq() :: migration_info()
   def create_func_eq do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:eq, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: compare_declarations(),
@@ -543,9 +524,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates `rational.__private__neq(a, b)` that backs the `<>` operator.
   """
-  @spec create_func_neq() :: {raw_sql(), raw_sql()}
+  @spec create_func_neq() :: migration_info()
   def create_func_neq do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:neq, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: compare_declarations(),
@@ -560,9 +541,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates `rational.__private__lt(a, b)` that backs the `<` operator.
   """
-  @spec create_func_lt() :: {raw_sql(), raw_sql()}
+  @spec create_func_lt() :: migration_info()
   def create_func_lt do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:lt, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: compare_declarations(),
@@ -577,9 +558,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates `rational.__private__lte(a, b)` that backs the `<=` operator.
   """
-  @spec create_func_lte() :: {raw_sql(), raw_sql()}
+  @spec create_func_lte() :: migration_info()
   def create_func_lte do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:lte, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: compare_declarations(),
@@ -594,9 +575,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates `rational.__private__gt(a, b)` that backs the `>` operator.
   """
-  @spec create_func_gt() :: {raw_sql(), raw_sql()}
+  @spec create_func_gt() :: migration_info()
   def create_func_gt do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:gt, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: compare_declarations(),
@@ -611,9 +592,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates `rational.__private__gte(a, b)` that backs the `>=` operator.
   """
-  @spec create_func_gte() :: {raw_sql(), raw_sql()}
+  @spec create_func_gte() :: migration_info()
   def create_func_gte do
-    Postgres.Utils.create_plpgsql_function(
+    PgTypeMigration.create_plpgsql_function(
       private_function(:gte, Migration.repo()),
       args: [a: :rational, b: :rational],
       declares: compare_declarations(),
@@ -632,9 +613,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
 
   Returns the absolute value of the input.
   """
-  @spec create_op_abs() :: {raw_sql(), raw_sql()}
+  @spec create_op_abs() :: migration_info()
   def create_op_abs do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :@,
       nil,
       :rational,
@@ -648,9 +629,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
 
   Flips the sign of `value`. Equivalent to `value * -1`.
   """
-  @spec create_op_minus() :: {raw_sql(), raw_sql()}
+  @spec create_op_minus() :: migration_info()
   def create_op_minus do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :-,
       nil,
       :rational,
@@ -662,9 +643,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `+` operator.
   """
-  @spec create_op_add() :: {raw_sql(), raw_sql()}
+  @spec create_op_add() :: migration_info()
   def create_op_add do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :+,
       :rational,
       :rational,
@@ -677,9 +658,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `-` operator.
   """
-  @spec create_op_sub() :: {raw_sql(), raw_sql()}
+  @spec create_op_sub() :: migration_info()
   def create_op_sub do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :-,
       :rational,
       :rational,
@@ -691,9 +672,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `*` operator.
   """
-  @spec create_op_mult() :: {raw_sql(), raw_sql()}
+  @spec create_op_mult() :: migration_info()
   def create_op_mult do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :*,
       :rational,
       :rational,
@@ -706,9 +687,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `/` operator.
   """
-  @spec create_op_div() :: {raw_sql(), raw_sql()}
+  @spec create_op_div() :: migration_info()
   def create_op_div do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :/,
       :rational,
       :rational,
@@ -720,9 +701,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `%` operator.
   """
-  @spec create_op_modulo() :: {raw_sql(), raw_sql()}
+  @spec create_op_modulo() :: migration_info()
   def create_op_modulo do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :%,
       :rational,
       :rational,
@@ -736,9 +717,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `=` operator.
   """
-  @spec create_op_eq() :: {raw_sql(), raw_sql()}
+  @spec create_op_eq() :: migration_info()
   def create_op_eq do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :=,
       :rational,
       :rational,
@@ -752,9 +733,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `<>` operator.
   """
-  @spec create_op_neq() :: {raw_sql(), raw_sql()}
+  @spec create_op_neq() :: migration_info()
   def create_op_neq do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :<>,
       :rational,
       :rational,
@@ -768,9 +749,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `<` operator.
   """
-  @spec create_op_lt() :: {raw_sql(), raw_sql()}
+  @spec create_op_lt() :: migration_info()
   def create_op_lt do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :<,
       :rational,
       :rational,
@@ -784,9 +765,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `<` operator.
   """
-  @spec create_op_lte() :: {raw_sql(), raw_sql()}
+  @spec create_op_lte() :: migration_info()
   def create_op_lte do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :<=,
       :rational,
       :rational,
@@ -800,9 +781,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `<` operator.
   """
-  @spec create_op_gt() :: {raw_sql(), raw_sql()}
+  @spec create_op_gt() :: migration_info()
   def create_op_gt do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :>,
       :rational,
       :rational,
@@ -816,9 +797,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a custom :rational, :rational `<` operator.
   """
-  @spec create_op_gte() :: {raw_sql(), raw_sql()}
+  @spec create_op_gte() :: migration_info()
   def create_op_gte do
-    Postgres.Utils.create_operator(
+    PgTypeMigration.create_operator(
       :>=,
       :rational,
       :rational,
@@ -834,9 +815,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   @doc """
   Creates a B-tree operator class to support indexing on comparison operations.
   """
-  @spec create_op_class_btree() :: {raw_sql(), raw_sql()}
+  @spec create_op_class_btree() :: migration_info()
   def create_op_class_btree do
-    Postgres.Utils.create_operator_class(
+    PgTypeMigration.create_operator_class(
       :rational_ops_btree,
       :rational,
       :btree,
@@ -863,9 +844,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   rational AS double precision
   ```
   """
-  @spec create_cast_double_precision() :: {raw_sql(), raw_sql()}
+  @spec create_cast_double_precision() :: migration_info()
   def create_cast_double_precision do
-    Postgres.Utils.create_cast(
+    PgTypeMigration.create_cast(
       :rational,
       :"double precision",
       private_function(:cast_to_double, Migration.repo())
@@ -880,9 +861,9 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
   bigint AS rational
   ```
   """
-  @spec create_cast_bigint_to_rational() :: {raw_sql(), raw_sql()}
+  @spec create_cast_bigint_to_rational() :: migration_info()
   def create_cast_bigint_to_rational do
-    Postgres.Utils.create_cast(
+    PgTypeMigration.create_cast(
       :bigint,
       :rational,
       private_function(:cast_bigint_to_rational, Migration.repo()),
@@ -930,25 +911,8 @@ defpgmodule Vtc.Ecto.Postgres.PgRational.Migrations do
     :ok
   end
 
-  @doc """
-  Returns the config-qualified name of the function for this type.
-  """
-  @spec function(atom(), Ecto.Repo.t()) :: String.t()
-  def function(name, repo) do
-    function_prefix = Postgres.Utils.type_function_prefix(repo, :rational)
-    "#{function_prefix}#{name}"
-  end
-
-  # Returns the config-qualified name of the function for this type.
-  @doc false
-  @spec private_function(atom(), Ecto.Repo.t()) :: String.t()
-  def private_function(name, repo) do
-    function_prefix = Postgres.Utils.type_private_function_prefix(repo, :rational)
-    "#{function_prefix}#{name}"
-  end
-
   # Returns declaration list for comparison operators.
-  @spec compare_declarations() :: Postgres.Utils.function_declarations()
+  @spec compare_declarations() :: PgTypeMigration.function_declarations()
   defp compare_declarations do
     [
       a_cmp: {:bigint, "((a).numerator * (b).denominator)"},
