+++ lib/ecto/postgres/postgres_migrations.ex
@@ -8,6 +8,7 @@ defpgmodule Vtc.Ecto.Postgres.Migrations do
   alias Vtc.Ecto.Postgres.PgFramerate
   alias Vtc.Ecto.Postgres.PgFramestamp
   alias Vtc.Ecto.Postgres.PgRational
+  alias Vtc.Ecto.Postgres.PgTypeMigration
 
   @doc """
   Runs all migrations. Safe to run multiple times when updates are required.
@@ -24,17 +25,17 @@ defpgmodule Vtc.Ecto.Postgres.Migrations do
 
   ## Options
 
-  - `pg_rational_opts`: Options to pass to
-    [PgRational.Migrations.create_all/0](`Vtc.Ecto.Postgres.PgRational.Migrations.create_all/0`).
+  - `rational`: Options to pass to
+    [PgRational.Migrations.run/0](`Vtc.Ecto.Postgres.PgRational.Migrations.run/0`).
 
-  - `pg_framerate_opts`: Options to pass to
-    [PgFramerate.Migrations.create_all/0](`Vtc.Ecto.Postgres.PgFramerate.Migrations.create_all/0`).
+  - `framerate`: Options to pass to
+    [PgFramerate.Migrations.run/0](`Vtc.Ecto.Postgres.PgFramerate.Migrations.run/0`).
 
-  - `pg_framestamp_opts`: Options to pass to
-    [PgFramestamp.Migrations.create_all/0](`Vtc.Ecto.Postgres.PgFramestamp.Migrations.create_all/0`).
+  - `framestamp`: Options to pass to
+    [PgFramestamp.Migrations.run/0](`Vtc.Ecto.Postgres.PgFramestamp.Migrations.run/0`).
 
-  - `pg_framestamp_range_opts`: Options to pass to
-    [PgFramestamp.Rage.Migrations.create_all/0](`Vtc.Ecto.Postgres.PgFramestamp.Range.Migrations.create_all/0`).
+  - `framestamp_range`: Options to pass to
+    [PgFramestamp.Rage.Migrations.run/0](`Vtc.Ecto.Postgres.PgFramestamp.Range.Migrations.run/0`).
 
   > #### Required Permissions {: .warning}
   >
@@ -50,22 +51,21 @@ defpgmodule Vtc.Ecto.Postgres.Migrations do
   > not return correct results, and ranges with different upper/lower bound types will
   > not be comparable.
   """
-  @spec migrate(
-          pg_rational_opts: [include: Keyword.t(), exclude: Keyword.t()],
-          pg_framerate_opts: [include: Keyword.t(), exclude: Keyword.t()],
-          pg_framestamp_opts: [include: Keyword.t(), exclude: Keyword.t()],
-          pg_framestamp_range_opts: [include: Keyword.t(), exclude: Keyword.t()]
+  @spec run(
+          rational: [include: Keyword.t(atom()), exclude: Keyword.t(atom())],
+          framerate: [include: Keyword.t(atom()), exclude: Keyword.t(atom())],
+          framestamp: [include: Keyword.t(atom()), exclude: Keyword.t(atom())],
+          framestamp_range: [include: Keyword.t(atom()), exclude: Keyword.t(atom())]
         ) :: :ok
-  def migrate(opts \\ []) do
-    pg_rational_opts = Keyword.get(opts, :pg_rational_opts, [])
-    pg_framerate_opts = Keyword.get(opts, :pg_framerate_opts, [])
-    pg_framestamp_opts = Keyword.get(opts, :pg_framestamp_opts, [])
-    pg_framestamp_range_opts = Keyword.get(opts, :pg_framestamp_range_opts, [])
+  def run(opts \\ []) do
+    migrations = [
+      PgRational.Migrations,
+      PgFramerate.Migrations,
+      PgFramestamp.Migrations,
+      PgFramestamp.Range.Migrations
+    ]
 
-    PgRational.Migrations.create_all(pg_rational_opts)
-    PgFramerate.Migrations.create_all(pg_framerate_opts)
-    PgFramestamp.Migrations.create_all(pg_framestamp_opts)
-    PgFramestamp.Range.Migrations.create_all(pg_framestamp_range_opts)
+    PgTypeMigration.run_all(migrations, opts)
 
     :ok
   end
