+++ setup.cfg
@@ -1,5 +1,5 @@
 [version]
-target = 0.15
+target = 0.16
 release =
 
 [testing]
